--- source/server/Client.cpp
@@ -1,3 +1,0 @@
-#include "Client.h"
-
-Client::Client(): registered(false), errcode(0) {}
