+++ source/server/commands/commands/irccommand.h
@@ -4,10 +4,11 @@
 #include <vector>
 
 #include "server/commands/commands/irccommands.h"
-#include "server/Server.h"
 
 namespace ircserv
 {
+
+class IRCServer;
 
 class IRCCommand
 {
@@ -19,7 +20,7 @@ private:
     void Shutdown(void);
 
 public:
-    virtual bool ProcessCommand(Server *serv) = 0;
+    virtual bool ProcessCommand(IRCServer *serv) = 0;
 
 private:
     virtual bool ValidateArgs(/*serverclass */) = 0;
