+++ source/server/commands/commands/ircpartcommand.h
@@ -18,7 +18,7 @@ private:
     void Shutdown(void);
 
 public:
-    bool ProcessCommand(Server *serv);
+    bool ProcessCommand(IRCServer *serv);
 
 private:
     bool ValidateArgs(/*serverclass */);
