+++ source/server/commands/manager/irccommandsmanager.h
@@ -1,12 +1,12 @@
 #pragma once
 
-#include "utils/singleton.h"
-
 #include "server/commands/parsing/lexer/irclexer.h"
 #include "server/commands/parsing/parser/ircparser.h"
 
 namespace ircserv
 {
+
+class IRCServer;
 
 class IRCCommandsManager
 {
@@ -16,16 +16,12 @@ public:
 private:
     void Initialize(void);
     void Shutdown(void);
-    DECLARE_SIMPLE_SINGLETON(IRCCommandsManager);
 
 public:
-    void ProcessCommand(const std::string& message, Server *serv);
+    void ProcessCommand(const std::string& message, IRCServer *serv);
 private:
     IRCLexer m_Lexer;
     IRCParser m_Parser;
 };
-
-#define GetIRCCommandsManager() IRCCommandsManager::GetInstance()
-
 
 }
