+++ source/server/commands/parsing/parser/ircparser.cpp
@@ -5,6 +5,7 @@
 #include "server/commands/parsing/tokens/ircargtoken.h"
 #include "server/commands/parsing/tokens/ircprefixtoken.h"
 
+
 namespace ircserv
 {
