+++ source/server/ircclient.cpp
@@ -1,0 +1,16 @@
+#include "main/precomp.h"
+
+#include "server/ircclient.h"
+
+namespace ircserv
+{
+
+IRCClient::IRCClient() : errcode(0), registered(false) 
+{
+}
+
+IRCClient::~IRCClient()
+{
+}
+
+}
